Move loadimages under lib/
